Converted bounding boxes to double from float.

+++ include/mps_voxels/octree_utils.h
@@ -44,8 +44,8 @@ public:
 
 	// min and max specified in camera frame
 	void completeShape(
-		const Eigen::Vector3f& min,
-		const Eigen::Vector3f& max,
+		const Eigen::Vector3d& min,
+		const Eigen::Vector3d& max,
 		const Eigen::Affine3f& worldTcamera,
 		octomap::OcTree* octree,
 		octomap::OcTree* subtree,
@@ -60,8 +60,8 @@ std::pair<octomap::point3d_collection, std::shared_ptr<octomap::OcTree>> getOccl
 	octomap::OcTree* octree,
 	const image_geometry::PinholeCameraModel& cameraModel,
 	const Eigen::Affine3d& cameraTworld,
-	const Eigen::Vector3f& minExtent,
-	const Eigen::Vector3f& maxExtent);
+	const Eigen::Vector3d& minExtent,
+	const Eigen::Vector3d& maxExtent);
 
 visualization_msgs::MarkerArray visualizeOctree(octomap::OcTree* tree, const std::string& globalFrame = "world", const std_msgs::ColorRGBA* base_color = nullptr);
